requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
crispy-bootstrap5==2025.6
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
Django==5.2.4
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.1
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-tables2==2.7.5
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fonttools==4.58.5
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonfield==3.2.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.1
packaging==25.0
pillow==11.3.0
preshed==3.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
randomcolor==0.4.4.6
requests==2.32.4
rich==14.0.0
scikit-learn==1.7.2
scipy==1.16.1
setuptools==80.9.0
shellingham==1.5.4
simpy==4.1.1
six==1.17.0
smart_open==7.3.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
swapper==1.4.0
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.11.0
wrapt==1.17.2
zopfli==0.2.3.post1
